QVAC-18064 fix: extract sanitizePrintableAscii, consolidate JSDoc
Extract sanitizePrintableAscii from anonymous namespaces in
TranslationModel.cpp and NmtLazyInitializeBackend.cpp into shared
nmt_utils.hpp/cpp. Consolidate duplicate JSDoc blocks on
getActiveBackendDescription in marian.js.
