+++ packages/qvac-lib-infer-nmtcpp/addon/src/model-interface/TranslationModel.cpp
@@ -14,20 +14,6 @@
 #include "nmt_utils.hpp"
 #include "qvac-lib-inference-addon-cpp/Logger.hpp"
 
-namespace {
-
-std::string sanitizePrintableAscii(const std::string& input) {
-  std::string out;
-  out.reserve(input.size());
-  for (char raw : input) {
-    unsigned char c = static_cast<unsigned char>(raw);
-    out.push_back((c >= 0x20 && c < 0x7F) ? static_cast<char>(c) : '?');
-  }
-  return out;
-}
-
-} // namespace
-
 namespace qvac_lib_inference_addon_nmt {
 
 std::string TranslationModel::getName() const {
