+++ packages/qvac-lib-infer-nmtcpp/marian.js
@@ -125,12 +125,12 @@ class TranslationInterface {
    * Returns the human-readable device description for the active GPU backend
    * (e.g. 'NVIDIA GeForce RTX 5070', 'Intel(R) UHD Graphics').
    * Returns '' when no GPU backend is loaded or model is unloaded.
-   * @returns {string}
-   */
-  /**
-   * Returns the active backend description, or '' if unavailable.
+   *
    * Silently catches native errors — the description is informational and
-   * callers should not fail when the backend cannot provide one.
+   * callers should not fail when the backend cannot provide one. This
+   * intentionally diverges from the error-throwing pattern used by other
+   * binding wrappers in this class.
+   * @returns {string}
    */
   getActiveBackendDescription () {
     if (this._handle === null) {
